requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-chess>=1.999
openai>=1.0.0
stockfish>=3.27.0
python-dotenv>=1.0.1
langsmith>=0.7.36
gitpython>=3.1.47